feat: add exception handling

+++ src/main/java/br/com/guilchaves/desafio/services/ClientService.java
@@ -3,13 +3,21 @@ package br.com.guilchaves.desafio.services;
 import br.com.guilchaves.desafio.dto.ClientDTO;
 import br.com.guilchaves.desafio.entities.Client;
 import br.com.guilchaves.desafio.repositories.ClientRepository;
+import br.com.guilchaves.desafio.services.exceptions.DatabaseException;
+import br.com.guilchaves.desafio.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
+
+//TODO
+// - Inserção de cliente retorna 422 e mensagens customizadas com dados inválidos
+// - Atualização de cliente retorna 422 e mensagens customizadas com dados inválidos
 
 @Service
 public class ClientService {
@@ -18,19 +26,20 @@ public class ClientService {
     private ClientRepository repository;
 
     @Transactional(readOnly = true)
-    public Page<ClientDTO> findAll(Pageable pageable){
+    public Page<ClientDTO> findAll(Pageable pageable) {
         Page<Client> result = repository.findAll(pageable);
         return result.map(ClientDTO::new);
     }
 
     @Transactional(readOnly = true)
-    public ClientDTO findById(Long id){
-        Client client = repository.findById(id).get();
+    public ClientDTO findById(Long id) {
+        Client client = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Resource not found"));
         return new ClientDTO(client);
     }
 
     @Transactional
-    public ClientDTO insert(ClientDTO dto){
+    public ClientDTO insert(ClientDTO dto) {
         Client entity = new Client();
         copyPropertiesToEntity(dto, entity);
         entity = repository.save(entity);
@@ -38,16 +47,30 @@ public class ClientService {
     }
 
     @Transactional
-    public ClientDTO update(Long id, ClientDTO dto){
-        Client entity = repository.getReferenceById(id);
-        copyPropertiesToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+    public ClientDTO update(Long id, ClientDTO dto) {
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyPropertiesToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
+
     }
 
-    @Transactional
-    public void delete(Long id){
-        repository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Resource not found");
+        }
+
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Data integrity violation: Unable to delete the resource with ID " + id +
+                    " due to existing related entities.");
+        }
     }
 
     private void copyPropertiesToEntity(ClientDTO dto, Client entity) {
